cloud-functions/publish/requirements.txt
requests==2.28.1
google-cloud-pubsub==2.13.11